tests/wpt/web-platform-tests/css/css-flexbox/canvas-dynamic-change-001.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>Canvas size dynamic change in flexbox layout</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#flex-items">
<meta name="assert" content="This test ensures proper layouting of canvas element as flex-items"/>
<link href="support/flexbox.css" rel="stylesheet">
<style>
.red {
  height: 400px;
  background: red;
}
canvas {
  background: green;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<div id=log></div>

<div class="red">
  <div class="flexbox column">
    <canvas id="canvas" data-expected-height="400"></canvas>
  </div>
</div>

<script>
var canvas = document.getElementById('canvas');

requestAnimationFrame(function() {
  canvas.height = 400;
  checkLayout('.flexbox');
});

</script>
